# packages/courses-backend/public/swagger.json
{
  "components": {
    "examples": {},
    "headers": {},
    "parameters": {},
    "requestBodies": {},
    "responses": {},
    "schemas": {
      "UserCredentials": {
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["username", "password"],
        "type": "object",
        "additionalProperties": false
      },
      "Course": {
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "number",
            "format": "double"
          },
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": ["title"],
        "type": "object",
        "additionalProperties": false
      },
      "ErrorResponse": {
        "properties": {
          "errorMessage": {
            "type": "string"
          }
        },
        "required": ["errorMessage"],
        "type": "object",
        "additionalProperties": false
      }
    },
    "securitySchemes": {}
  },
  "info": {
    "title": "backend",
    "version": "1.0.0",
    "license": {
      "name": "ISC"
    },
    "contact": {
      "name": "vasiliev-alexey"
    }
  },
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "RegisterNewUser",
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Register Failed",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCredentials"
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "Login",
        "responses": {
          "204": {
            "description": "No content"
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCredentials"
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "operationId": "Logout",
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [],
        "parameters": []
      }
    },
    "/courses": {
      "get": {
        "operationId": "GetCourses",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Course"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": []
      },
      "post": {
        "operationId": "CreateNewCourse",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          },
          "400": {
            "description": "Unauthorized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Course"
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}